test: package-private 변경에 따른 에러 재활용
